requirements.txt
dj-database-url==0.5.0
Django==3.2.4
django-on-heroku==1.1.2
gunicorn==20.1.0
imageio==2.9.0
keras==2.6.0
matplotlib==3.4.2
numpy==1.19.5
tensorflow-cpu==2.6.0
tensorflow-hub==0.12.0
asgiref==3.3.4
django-heroku==0.3.1
psycopg2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise
boto3
django-storages